fix(terraform,helmfile): declare plan/diff outputs  (#33)
For the outputs to be exposed to the outside they must be declared in the action.yml file.

As part of this change I am renaming them to be more specific (e.g. output -> plan-output) to:

allow enable different descriptions per subcommand that generates the action
avoiding overloading the meaning if a more generic name is used
I'm also removing the outputs generated by terraform init, fmt, and validate, since we were not using them anywhere. We can add them back once we identify a real use case. This is not a breaking change since those outputs where never exposed from the action anyway.