Pass through arguments to terraform sub commands
Turns out that when we were doing this switch statement, we weren't
parsing for the command and then passing on any additional flags.

In later versions of Terraform, the `-lock-timeout` behavior changes to
immediately fail if there's an existing lock instead of waiting for the
lock to be released as the default behavior. When I tried to set
`-lock-timeout`, the commands stopped matching in this switch block.